Move all rest related files around
