+++ src/web/middleware/require-permissions.ts
