+++ src/web/rest/routes/bikes/index.ts
