+++ src/web/rest/routes/categories/categories.post.ts
@@ -1,9 +1,9 @@
 import * as Koa from "koa";
 import * as compose from "koa-compose";
 import {CreateCategoryRequestSchema} from "../../schema/categories";
-import * as Logic from "../../../web/logic/categories";
-import RequirePermission from "../../middleware/require-permissions";
-import {AuthLevel} from "../../../auth/role";
+import * as Logic from "../../../logic/categories";
+import RequirePermission from "../../../middleware/require-permissions";
+import {AuthLevel} from "../../../../auth/role";
 
 const PostCategories: Koa.Middleware = compose([
     RequirePermission(AuthLevel.Manager),
