+++ src/web/rest/routes/categories/index.ts
