+++ src/web/rest/routes/images/index.ts
