+++ src/web/rest/schema/bikes.ts
