+++ src/web/rest/schema/bookings.ts
