+++ src/web/rest/schema/categories.ts
