+++ src/web/rest/schema/images.ts
